Aggressive refactor - stop bringing Forms with us
Upon testing on a project with silverstripe/elemental installed and
added to UserDefinedForm (the page that times out on save) an error was
found in testing due insufficient information captured to be able to
recreate the form as required for the queued job to be able to complete
its processing.

It was considered that Forms are already their own hydrators as this is
how submissions work - a form builds itself and then handles the
submission. We would only need to handle the submission data and perhaps
some extra info in order to build the form... such as the current page
ID for CMSMain. But CMSMain fetches that from the submission URL and no
other source, complicating matters.

It then occurred to me that Form::saveInto is _unlikely_ to be the major
source of performance woes, even with some advanced "fake properties"
via `save{Name}()` methods, the likelihood of advanced performance
hitting logic being hidden in them seems low.

Taking this gamble I've elected to save the form into the object, and
then simply serialise the whole DataObject for writing at a later date.
This keeps untold complexity away from the module. If there is still a
speed issue with this solution found in the future we can deal with it
then.